feat: new info sequences library
