+++ .vscode/settings.json
@@ -2,6 +2,9 @@
     "files.associations": {
         "i2c.h": "c",
         "main.h": "c",
-        "stm32f4xx_hal.h": "c"
+        "stm32f4xx_hal.h": "c",
+        "stm32f4xx_hal_conf.h": "c",
+        "bno055.h": "c",
+        "testing.h": "c"
     }
 }
