+++ Core/Inc/main.h
@@ -32,6 +32,8 @@ extern "C" {
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
 
+#include "stm32f411xe.h"
+
 /* USER CODE END Includes */
 
 /* Exported types ------------------------------------------------------------*/
